Have bot responding on all channels and saving to the DB

+++ src/main/kotlin/ok/bot/bender/repository/BenderRepository.kt
@@ -3,6 +3,6 @@ package ok.bot.bender.repository
 import ok.bot.bender.entity.TransactionHistory
 import org.springframework.data.repository.CrudRepository
 
-interface BenderRepository : CrudRepository<TransactionHistory, Long> {
+interface BenderRepository : CrudRepository<TransactionHistory, Int> {
 
 }
